Improve the UX while uploading the code.
